+++ tests/testthat/test_modelsum.R
@@ -632,3 +632,15 @@ test_that("statistic.F works (#262)", {
     )
   )
 })
+
+test_that("Nevents works (#266)", {
+  tab3 <- modelsum(Surv(fu.time,fu.stat)~sex, data=mockstudy, survival.stats=c('HR','p.value','Nmiss','Nevents','N'), family = "survival")
+  expect_identical(
+    capture.kable(summary(tab3)),
+    c("|               |HR    |p.value |Nevents |N    |",
+      "|:--------------|:-----|:-------|:-------|:----|",
+      "|**sex Female** |1.002 |0.975   |1356    |1499 |"
+    )
+  )
+})
+